test: Aggregate SamRecordHandler tests in one class
