+++ server/catgenome/src/test/java/com/epam/catgenome/manager/bam/SamRecordHandlerTest.java
@@ -35,6 +35,7 @@ import com.epam.catgenome.entity.bam.BamQueryOption;
 import com.epam.catgenome.entity.bam.BamTrackMode;
 import com.epam.catgenome.entity.bam.TrackDirectionType;
 import com.epam.catgenome.entity.reference.Reference;
+import com.epam.catgenome.exception.SamAlignmentException;
 import com.epam.catgenome.manager.bam.filters.Filter;
 import com.epam.catgenome.manager.bam.filters.MiddleSAMRecordFilter;
 import com.epam.catgenome.manager.bam.handlers.SAMRecordHandler;
@@ -125,5 +126,18 @@ public class SamRecordHandlerTest extends AbstractManagerTest {
         recordHandler.add(rec2);
 
         Assert.assertEquals(2, recordHandler.getSifter().getFilteredReadsCount());
+    }
+
+    @Test(expected = SamAlignmentException.class)
+    @Transactional(propagation = Propagation.REQUIRES_NEW, rollbackFor = Exception.class)
+    public void matchBeyondReferenceTest() throws IOException {
+
+        recordHandler = new SAMRecordHandler(1, endTrack, referenceManager, filter, options);
+
+        final SAMRecordSetBuilder set = new SAMRecordSetBuilder();
+        set.setReadLength(readLength);
+        final SAMRecord rec = set.addFrag("read3", 0, 100, false, false, "75M", "*", 151);
+
+        recordHandler.add(rec);
     }
 }
